Added new Exception Handling.
