--- src/app/class/consonant.ts
@@ -1,18 +1,0 @@
-import { Manner, Onset, Place, Unicode } from './../enum/index';
-import { Sound } from './sound';
-
-export class Consonant extends Sound {
-  place: string;
-  manner: string;
-  voicing: boolean;
-  onset: Onset;
-
-  constructor(place: string, manner: string, voicing: boolean, ipa_unicode: string) {
-    super(ipa_unicode);
-    this.place = place;
-    this.manner = manner;
-    this.voicing = voicing;
-    this.onset = 0;
-  }
-
-}
