+++ src/app/class/index.ts
@@ -1,7 +1,13 @@
-export { Consonant } from './consonant';
-export { Gloss } from './gloss';
-export { Inventory } from './inventory';
-export { Parser } from './parser';
-export { Permuter } from './permuter';
-export { Sound } from './sound';
-export { Vowel } from './vowel';
+export { Consonant } from './ipa/consonant';
+export { EmptyInputError } from './error/empty-input-error';
+export { GenericError } from './error/generic-error';
+export { Gloss } from './ipa/gloss';
+export { Inventory } from './ipa/inventory';
+export { NullArgumentError } from './error/null-argument-error';
+export { Parser } from './ipa/parser';
+export { Permuter } from './ipa/permuter';
+export { Sound } from './ipa/sound';
+export { Vowel } from './ipa/vowel';
+export { WordStructure } from './ipa/word-structure';
+export { WordStructureComparer } from './ipa/word-structure-comparer';
+export { WordStructureComponent } from './ipa/word-structure-component';
