+++ src/app/class/ipa/word-structure-comparer.ts
