+++ src/app/class/ipa/word-structure-component.ts
