--- src/app/class/sound.ts
@@ -1,9 +1,0 @@
-export class Sound {
-  ipa_unicode: string;
-  isSelected: boolean;
-
-  constructor(ipa_unicode: string) {
-    this.ipa_unicode = ipa_unicode;
-    this.isSelected = false;
-  }
-}
