+++ src/app/component/error/error.component.html
@@ -1,0 +1,3 @@
+<p>
+  error works!
+</p>
