+++ src/app/component/index.ts
@@ -1,3 +1,4 @@
+export { ErrorComponent } from './error/error.component';
 export { IpaComponent } from './ipa/ipa.component';
 export { InputComponent } from './input/input.component';
 export { OutputComponent } from './output/output.component';
