+++ src/app/enum/parser-symbols.enum.ts
@@ -1,6 +1,4 @@
 export enum ParserSymbols {
-
   OPTIONAL_START = '(',
   OPTIONAL_END = ')'
-
 }
