+++ src/app/enum/sound-symbols.enum.ts
@@ -1,6 +1,4 @@
 export enum SoundSymbols {
-
   VOWEL = 'V',
   CONSONANT = 'C'
-
 }
