+++ src/app/service/error.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class ErrorService {
+
+  constructor() { }
+
+  public displayError(message: string) {
+
+  }
+
+}
