+++ src/app/service/index.ts
@@ -1,2 +1,3 @@
+export { ErrorService } from './error.service';
 export { IpaService } from './ipa.service';
 export { TranscriptionService } from './transcription.service';
