+++ src/app/test/empty-inventory-error.ts
@@ -1,0 +1,2 @@
+export class EmptyInventoryError {
+}
