+++ src/app/test/error.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { ErrorService } from './error.service';
+
+describe('ErrorService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [ErrorService]
+    });
+  });
+
+  it('should be created', inject([ErrorService], (service: ErrorService) => {
+    expect(service).toBeTruthy();
+  }));
+});
